Move timeout methods into shared class
